Moved location of type SynthesizerBlockInfo
