+++ packages/frontend/synthesizer/src/interface/index.ts
@@ -1,2 +1,2 @@
 export { mapToStr } from './debugging/utils.ts'
-export { createSynthesizerOptsForSimulationFromRPC, type SynthesizerSimulationOpts } from './rpc/rpc.ts'
+export { createSynthesizerOptsForSimulationFromRPC, type SynthesizerSimulationOpts, type SynthesizerBlockInfo } from './rpc/index.ts'
