+++ packages/frontend/synthesizer/src/interface/rpc/index.ts
@@ -1,0 +1,2 @@
+export * from './types.ts'
+export * from './rpc.ts'
