+++ packages/frontend/synthesizer/src/interface/rpc/types.ts
@@ -1,0 +1,11 @@
+export type SynthesizerBlockInfo = {
+  coinBase: `0x${string}`,
+  timeStamp: `0x${string}`,
+  blockNumber: `0x${string}`,
+  prevRanDao: `0x${string}`,
+  gasLimit: `0x${string}`,
+  chainId: `0x${string}`,
+  selfBalance: `0x${string}`,
+  baseFee: `0x${string}`,
+  prevBlockHashes: `0x${string}`[],
+}
